[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17720 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,4 +1,3 @@
-
 import java.util.Scanner;
 
 public class Main {
@@ -6,8 +5,8 @@ public class Main {
 	public static void main(String[] args) {
 		
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		long a = sc.nextInt();
+		long b = sc.nextInt();
 		
 		System.out.println((a+b)*(a-b));
